refactored stacking
